Split compact testing into multiple modules
- Required because we can only run the tests when the appropiate testing
  flag is enabled.
