+++ trace_decoder/src/compact/compact_processing_common.rs
@@ -1009,251 +1009,3 @@ pub struct ProcessedCompactOutput<T> {
     pub witness_out: T,
 }
 
-#[cfg(test)]
-mod tests {
-    use mpt_trie::nibbles::Nibbles;
-
-    use super::Instruction;
-    use crate::compact::{
-        compact_debug_tools::parse_just_to_instructions,
-        compact_processing_common::key_bytes_to_nibbles,
-        complex_test_payloads::{
-            MPT_TEST_PAYLOAD_1, MPT_TEST_PAYLOAD_2, MPT_TEST_PAYLOAD_3, MPT_TEST_PAYLOAD_4,
-            MPT_TEST_PAYLOAD_5, MPT_TEST_PAYLOAD_6, SMT_TEST_PAYLOAD_1, SMT_TEST_PAYLOAD_10,
-            SMT_TEST_PAYLOAD_11, SMT_TEST_PAYLOAD_12, SMT_TEST_PAYLOAD_13, SMT_TEST_PAYLOAD_14,
-            SMT_TEST_PAYLOAD_15, SMT_TEST_PAYLOAD_16, SMT_TEST_PAYLOAD_17, SMT_TEST_PAYLOAD_18,
-            SMT_TEST_PAYLOAD_19, SMT_TEST_PAYLOAD_2, SMT_TEST_PAYLOAD_20, SMT_TEST_PAYLOAD_21,
-            SMT_TEST_PAYLOAD_22, SMT_TEST_PAYLOAD_23, SMT_TEST_PAYLOAD_24, SMT_TEST_PAYLOAD_25,
-            SMT_TEST_PAYLOAD_3, SMT_TEST_PAYLOAD_4, SMT_TEST_PAYLOAD_5, SMT_TEST_PAYLOAD_6,
-            SMT_TEST_PAYLOAD_7, SMT_TEST_PAYLOAD_8, SMT_TEST_PAYLOAD_9,
-        },
-    };
-
-    const SIMPLE_PAYLOAD_STR: &str = "01004110443132333400411044313233340218300042035044313233350218180158200000000000000000000000000000000000000000000000000000000000000012";
-
-    fn init() {
-        let _ = pretty_env_logger::try_init();
-    }
-
-    fn h_decode_key(h_bytes: &str) -> Nibbles {
-        let bytes = hex::decode(h_bytes).unwrap();
-        key_bytes_to_nibbles(&bytes)
-    }
-
-    fn h_decode(b_str: &str) -> Vec<u8> {
-        hex::decode(b_str).unwrap()
-    }
-
-    #[test]
-    fn simple_instructions_are_parsed_correctly() {
-        init();
-
-        let bytes = hex::decode(SIMPLE_PAYLOAD_STR).unwrap();
-        let instrs = parse_just_to_instructions(bytes);
-
-        let instrs = match instrs {
-            Ok(x) => x,
-            Err(err) => panic!("{}", err),
-        };
-
-        let expected_instrs = vec![
-            Instruction::Leaf(h_decode_key("10"), h_decode("31323334")),
-            Instruction::Leaf(h_decode_key("10"), h_decode("31323334")),
-            Instruction::Branch(0b00110000),
-            Instruction::Leaf(h_decode_key("0350"), h_decode("31323335")),
-            Instruction::Branch(0b00011000),
-            Instruction::Extension(h_decode_key(
-                "0000000000000000000000000000000000000000000000000000000000000012",
-            )),
-        ];
-
-        for (i, expected_instr) in expected_instrs.into_iter().enumerate() {
-            assert_eq!(expected_instr, instrs[i])
-        }
-    }
-
-    #[test]
-    fn mpt_complex_payload_1() {
-        init();
-        MPT_TEST_PAYLOAD_1.parse_and_check_hash_matches_with_debug();
-    }
-
-    #[test]
-    fn mpt_complex_payload_2() {
-        init();
-        MPT_TEST_PAYLOAD_2.parse_and_check_hash_matches_with_debug();
-    }
-
-    #[test]
-    fn mpt_complex_payload_3() {
-        init();
-        MPT_TEST_PAYLOAD_3.parse_and_check_hash_matches_with_debug();
-    }
-
-    #[test]
-    fn mpt_complex_payload_4() {
-        init();
-        MPT_TEST_PAYLOAD_4.parse_and_check_hash_matches_with_debug();
-    }
-
-    #[test]
-    fn mpt_complex_payload_5() {
-        init();
-        MPT_TEST_PAYLOAD_5.parse_and_check_hash_matches_with_debug();
-    }
-
-    #[test]
-    fn mpt_complex_payload_6() {
-        init();
-        MPT_TEST_PAYLOAD_6.parse_and_check_hash_matches_with_debug();
-    }
-
-    #[test]
-    fn smt_complex_payload_1() {
-        init();
-        SMT_TEST_PAYLOAD_1.parse_and_check_hash_matches_with_debug_smt();
-    }
-
-    #[test]
-    fn smt_complex_payload_2() {
-        init();
-        SMT_TEST_PAYLOAD_2.parse_and_check_hash_matches_with_debug_smt();
-    }
-
-    #[test]
-    fn smt_complex_payload_3() {
-        init();
-        SMT_TEST_PAYLOAD_3.parse_and_check_hash_matches_with_debug_smt();
-    }
-
-    #[test]
-    fn smt_complex_payload_4() {
-        init();
-        SMT_TEST_PAYLOAD_4.parse_and_check_hash_matches_with_debug_smt();
-    }
-
-    #[test]
-    fn smt_complex_payload_5() {
-        init();
-        SMT_TEST_PAYLOAD_5.parse_and_check_hash_matches_with_debug_smt();
-    }
-
-    #[test]
-    fn smt_complex_payload_6() {
-        init();
-        SMT_TEST_PAYLOAD_6.parse_and_check_hash_matches_with_debug_smt();
-    }
-
-    #[test]
-    fn smt_complex_payload_7() {
-        init();
-        SMT_TEST_PAYLOAD_7.parse_and_check_hash_matches_with_debug_smt();
-    }
-
-    #[test]
-    fn smt_complex_payload_8() {
-        init();
-        SMT_TEST_PAYLOAD_8.parse_and_check_hash_matches_with_debug_smt();
-    }
-
-    #[test]
-    fn smt_complex_payload_9() {
-        init();
-        SMT_TEST_PAYLOAD_9.parse_and_check_hash_matches_with_debug_smt();
-    }
-
-    #[test]
-    fn smt_complex_payload_10() {
-        init();
-        SMT_TEST_PAYLOAD_10.parse_and_check_hash_matches_with_debug_smt();
-    }
-
-    #[test]
-    fn smt_complex_payload_11() {
-        init();
-        SMT_TEST_PAYLOAD_11.parse_and_check_hash_matches_with_debug_smt();
-    }
-
-    #[test]
-    fn smt_complex_payload_12() {
-        init();
-        SMT_TEST_PAYLOAD_12.parse_and_check_hash_matches_with_debug_smt();
-    }
-    #[test]
-    fn smt_complex_payload_13() {
-        init();
-        SMT_TEST_PAYLOAD_13.parse_and_check_hash_matches_with_debug_smt();
-    }
-
-    #[test]
-    fn smt_complex_payload_14() {
-        init();
-        SMT_TEST_PAYLOAD_14.parse_and_check_hash_matches_with_debug_smt();
-    }
-
-    #[test]
-    fn smt_complex_payload_15() {
-        init();
-        SMT_TEST_PAYLOAD_15.parse_and_check_hash_matches_with_debug_smt();
-    }
-
-    #[test]
-    fn smt_complex_payload_16() {
-        init();
-        SMT_TEST_PAYLOAD_16.parse_and_check_hash_matches_with_debug_smt();
-    }
-
-    #[test]
-    fn smt_complex_payload_17() {
-        init();
-        SMT_TEST_PAYLOAD_17.parse_and_check_hash_matches_with_debug_smt();
-    }
-
-    #[test]
-    fn smt_complex_payload_18() {
-        init();
-        SMT_TEST_PAYLOAD_18.parse_and_check_hash_matches_with_debug_smt();
-    }
-
-    #[test]
-    fn smt_complex_payload_19() {
-        init();
-        SMT_TEST_PAYLOAD_19.parse_and_check_hash_matches_with_debug_smt();
-    }
-
-    #[test]
-    fn smt_complex_payload_20() {
-        init();
-        SMT_TEST_PAYLOAD_20.parse_and_check_hash_matches_with_debug_smt();
-    }
-
-    #[test]
-    fn smt_complex_payload_21() {
-        init();
-        SMT_TEST_PAYLOAD_21.parse_and_check_hash_matches_with_debug_smt();
-    }
-
-    #[test]
-    fn smt_complex_payload_22() {
-        init();
-        SMT_TEST_PAYLOAD_22.parse_and_check_hash_matches_with_debug_smt();
-    }
-
-    #[test]
-    fn smt_complex_payload_23() {
-        init();
-        SMT_TEST_PAYLOAD_23.parse_and_check_hash_matches_with_debug_smt();
-    }
-
-    #[test]
-    fn smt_complex_payload_24() {
-        init();
-        SMT_TEST_PAYLOAD_24.parse_and_check_hash_matches_with_debug_smt();
-    }
-
-    #[test]
-    fn smt_complex_payload_25() {
-        init();
-        SMT_TEST_PAYLOAD_25.parse_and_check_hash_matches_with_debug_smt();
-    }
-}
