+++ trace_decoder/src/compact/complex_test_payloads.rs
@@ -3,11 +3,7 @@ use mpt_trie::partial_trie::PartialTrie;
 use smt_trie::utils::hashout2u;
 
 use super::{
-    compact_mpt_processing::MptPreImageProcessing,
-    compact_processing_common::{CompactParsingResult, Header, ProcessedCompactOutput},
-    compact_smt_processing::SmtPreImageProcessing,
-    compact_to_mpt_trie::StateTrieExtractionOutput,
-    compact_to_smt_trie::SmtStateTrieExtractionOutput,
+    compact_processing_common::{key_bytes_to_nibbles, CompactParsingResult, Header, ProcessedCompactOutput},
 };
 use crate::{
     aliased_crate_types::AccountRlp,
@@ -16,230 +12,34 @@ use crate::{
     utils::{print_value_and_hash_nodes_of_storage_trie, print_value_and_hash_nodes_of_trie},
 };
 
-pub(crate) const MPT_TEST_PAYLOAD_1: TestProtocolInputAndRoot = TestProtocolInputAndRoot { byte_str: "01055821033601462093b5945d1676df093446790fd31b20e7b12a2e8e5e09d068109616b0084a021e19e0c9bab240000005582103468288056310c82aa4c01a7e12a10f8111a0560e72b700555479031b86c357d0084101031a697e814758281972fcd13bc9707dbcd2f195986b05463d7b78426508445a0405582103b70e80538acdabd6137353b0f9d8d149f4dba91e8be2e7946e409bfdbe685b900841010558210389802d6ed1a28b049e9d4fe5334c5902fd9bc00c42821c82f82ee2da10be90800841010558200256274a27dd7524955417c11ecd917251cc7c4c8310f4c7e4bd3c304d3d9a79084a021e19e0c9bab2400000055820023ab0970b73895b8c9959bae685c3a19f45eb5ad89d42b52a340ec4ac204d190841010219102005582103876da518a393dbd067dc72abfa08d475ed6447fca96d92ec3f9e7eba503ca6100841010558210352688a8f926c816ca1e079067caba944f158e764817b83fc43594370ca9cf62008410105582103690b239ba3aaf993e443ae14aeffc44cf8d9931a79baed9fa141d0e4506e131008410102196573", root_str: "6a0673c691edfa4c4528323986bb43c579316f436ff6f8b4ac70854bbd95340b" };
+pub(super) type ProcessedCompactPrestateFn<T, U> = fn(T) -> CompactParsingResult<ProcessedCompactOutput<U>>;
 
-pub(crate) const MPT_TEST_PAYLOAD_2: TestProtocolInputAndRoot = TestProtocolInputAndRoot { byte_str: "01055821033601462093b5945d1676df093446790fd31b20e7b12a2e8e5e09d068109616b0084a021e19e0c9bab240000005582103468288056310c82aa4c01a7e12a10f8111a0560e72b700555479031b86c357d0084101031a697e814758281972fcd13bc9707dbcd2f195986b05463d7b78426508445a0405582103b70e80538acdabd6137353b0f9d8d149f4dba91e8be2e7946e409bfdbe685b900841010558210389802d6ed1a28b049e9d4fe5334c5902fd9bc00c42821c82f82ee2da10be90800841010558200256274a27dd7524955417c11ecd917251cc7c4c8310f4c7e4bd3c304d3d9a790c014a021e0c000250c782fa00055820023ab0970b73895b8c9959bae685c3a19f45eb5ad89d42b52a340ec4ac204d1908410102191020055820021eec2b84f0ba344fd4b4d2f022469febe7a772c4789acfc119eb558ab1da3d08480de0b6b3a76400000558200276da518a393dbd067dc72abfa08d475ed6447fca96d92ec3f9e7eba503ca61084101021901200558210352688a8f926c816ca1e079067caba944f158e764817b83fc43594370ca9cf62008410105582103690b239ba3aaf993e443ae14aeffc44cf8d9931a79baed9fa141d0e4506e131008410102196573", root_str: "e779761e7f0cf4bb2b5e5a2ebac65406d3a7516d46798040803488825a01c19c" };
-
-pub(crate) const MPT_TEST_PAYLOAD_3: TestProtocolInputAndRoot = TestProtocolInputAndRoot { byte_str: "01055821033601462093b5945d1676df093446790fd31b20e7b12a2e8e5e09d068109616b0084a021e19e0c9bab240000005582103468288056310c82aa4c01a7e12a10f8111a0560e72b700555479031b86c357d0084101031a697e814758281972fcd13bc9707dbcd2f195986b05463d7b78426508445a0405582103b70e80538acdabd6137353b0f9d8d149f4dba91e8be2e7946e409bfdbe685b900841010558210389802d6ed1a28b049e9d4fe5334c5902fd9bc00c42821c82f82ee2da10be90800841010558200256274a27dd7524955417c11ecd917251cc7c4c8310f4c7e4bd3c304d3d9a790c024a021e0a9cae36fa8e4788055820023ab0970b73895b8c9959bae685c3a19f45eb5ad89d42b52a340ec4ac204d1908410102191020055820021eec2b84f0ba344fd4b4d2f022469febe7a772c4789acfc119eb558ab1da3d08480f43fc2c04ee00000558200276da518a393dbd067dc72abfa08d475ed6447fca96d92ec3f9e7eba503ca61084101021901200558210352688a8f926c816ca1e079067caba944f158e764817b83fc43594370ca9cf62008410105582103690b239ba3aaf993e443ae14aeffc44cf8d9931a79baed9fa141d0e4506e131008410102196573", root_str: "6978d65a3f2fc887408cc28dbb796836ff991af73c21ea74d03a11f6cdeb119c" };
-
-pub(crate) const MPT_TEST_PAYLOAD_4: TestProtocolInputAndRoot = TestProtocolInputAndRoot { byte_str: "0103a6885b3731702da62e8e4a8f584ac46a7f6822f4e2ba50fba902f67b1588d23b005821028015657e298d35290e69628be03d91f74d613caf3afdbe09138cfa415efe2f5044deadbeef0558210218b289936a0874cccee65712c88cdaa0a305b004d3fda2942b2b2dc54f14f6110b443b9aca0004", root_str: "69a5a7a8f99161a35e8b64975d8c6af10db4eee7bd956418839d8ff763aaf00c" };
-
-pub(crate) const MPT_TEST_PAYLOAD_5: TestProtocolInputAndRoot = TestProtocolInputAndRoot {
-    byte_str: include_str!("large_test_payloads/mpt_test_payload_5.txt"),
-    root_str: "2b5a703bdec53099c42d7575f8cd6db85d6f2226a04e98e966fcaef87868869b",
-};
-
-pub(crate) const MPT_TEST_PAYLOAD_6: TestProtocolInputAndRoot = TestProtocolInputAndRoot {
-    byte_str: include_str!("large_test_payloads/mpt_test_payload_6.txt"),
-    root_str: "135a0c66146c60d7f78049b3a3486aae3e155015db041a4650966e001f9ba301",
-};
-
-// TestSimpleSMTWitness, 1 node
-pub(crate) const SMT_TEST_PAYLOAD_1: TestProtocolInputAndRoot = TestProtocolInputAndRoot { byte_str: "0107035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000014101", root_str: "9d2ba51630afa706b806db84f4136d7b6a0df6bf37008aa9a2e70c8a99f161ee" };
-
-// TestSimpleSMTWitness, 50 nodes
-pub(crate) const SMT_TEST_PAYLOAD_2: TestProtocolInputAndRoot = TestProtocolInputAndRoot {
-    byte_str: "0102030203020302030203020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000008410807035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000004641460203020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000015411507035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000047414707035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000050415002020201020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000044414407035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000001841180203020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000013411307035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000384138020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000024412407035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000254125020302030203020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000027412707035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000000641060202020102020202020202020201020102010201020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000040414007035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000344134020302030202020102020201020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000049414907035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000023412307035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000294129020307035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000314131020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000021412102010201020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000035413507035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000004341430201020302020201020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000030413007035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000016411607035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000022412202030203020307035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000044104020302030201020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000020412007035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000041414107035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000011411107035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000094109020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000001410107035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000028412802030203020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000003410307035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000004841480203020202030201020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000032413207035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000005410507035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000014411402030203020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000012411207035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000036413607035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000000741070201020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000026412607035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000000241020203020307035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000003741370203020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000039413907035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000045414502010202020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000017411707035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000003341330203020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000010411007035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000019411907035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000424142",
-    root_str: "c3c20d5b5277a51aab01c464e7a7b4aeb2506ff2015f9f0d69f869fc1f723cab",
-};
-
-// TestSimpleSMTWitness, 100 nodes
-pub(crate) const SMT_TEST_PAYLOAD_3: TestProtocolInputAndRoot = TestProtocolInputAndRoot {
-    byte_str: "0102030203020302030203020307035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000084108020102030202020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000063416307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000046414607035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000008641860203020302020201020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000015411507035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000007041700201020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000057415707035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000047414707035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000050415002030203020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000098419807035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000099419907035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000008741870203020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000044414407035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000018411807035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000009041900203020302030203020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000095419507035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000091419107035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000013411307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000093419302020202020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000051415107035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000003841380203020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000024412407035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000644164020307035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000774177020202020202020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000072417207035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000025412502030203020302030201020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000027412707035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000058415807035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000064106020307035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000764176020302020202020202020201020102010201020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000040414007035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000003441340202020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000081418107035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000075417502030203020307035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000744174020102020201020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000049414907035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000023412307035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000294129020307035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000314131020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000021412102010201020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000035413507035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000004341430203020302020201020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000030413007035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000016411607035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000022412202030203020102010202020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000052415207035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000097419707035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000068416807035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000884188020302030203020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000082418207035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000044104020302030201020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000020412007035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000041414102020203020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000011411107035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000062416207035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000614161020302020201020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000067416707035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000604160020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000100420100020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000069416907035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000000941090203020307035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000014101020307035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000664166020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000079417907035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000005641560203020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000083418307035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000009641960202020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000055415507035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000028412802030203020302010203020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000071417107035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000003410307035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000008541850202020202020201020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000053415307035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000004841480203020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000059415902030203020307035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000324132020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000080418007035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000005410507035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000094419407035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000014411402030203020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000012411207035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000036413607035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000000741070201020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000026412607035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000024102020302030201020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000037413707035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000005441540203020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000039413907035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000045414502010202020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000017411707035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000003341330203020302030202020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000084418407035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000010411007035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000006541650201020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000019411907035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000924192020307035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000734173020307035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000004241420202020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000089418907035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000784178",
-    root_str: "47206520075c4d8410f6a77d64c655d00b7843ac28bad4d07feef6e7d8db1f32",
-};
-
-// TestSMTWitnessRetainList, 1 node
-pub(crate) const SMT_TEST_PAYLOAD_4: TestProtocolInputAndRoot = TestProtocolInputAndRoot {
-    byte_str: "0102030202020304440102030407025471dd1027069078091b3ca48093b00e4735b206245820fee777ac8b47f421d9f44ddd09c5efdd8ff00fbc7bbeb8169054d58a1b4b912207045471dd1027069078091b3ca48093b00e4735b206244104020302010201020307015471dd1027069078091b3ca48093b00e4735b20624410107005471dd1027069078091b3ca48093b00e4735b20624443b9aca0007035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000000544deadbeef",
-    root_str: "3390654ab1d7746a289e8530358e4e8f3474bec49816c740b469acfc431a244c",
-};
-
-// TestSMTWitnessRetainList, 50 node
-pub(crate) const SMT_TEST_PAYLOAD_5: TestProtocolInputAndRoot = TestProtocolInputAndRoot { byte_str: "010203020302030203020303fd6c3576caff9e650a709df487122958032121a9324f245d0588076e7d79ebca0201020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000015411507035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000474147034a0a8ade2d571c3bbafb46ebeaba9af81924e413fe1df89f6ede57203e5516c30203020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000013411307035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000384138035e922f930d52accaba7c46d0a97b2b309290d5385380bf8846acfb38aff55a4d020302030203020302010201020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000027412704440102030407025471dd1027069078091b3ca48093b00e4735b206245820fee777ac8b47f421d9f44ddd09c5efdd8ff00fbc7bbeb8169054d58a1b4b912207035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000000641060314bd622731b8bc10310784233b5d2549b4ba5bd6c70df619ac517314f3526bfb020303ebd30c9db62098e85b738370e9dfd0a7acba5fb656d6b5be4201f8dfbaaebe49020307035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000314131036311f89149d48b4e97cda99d7377396c15156dd7ddd3ee0f8eca56902e94ad7002010203020307045471dd1027069078091b3ca48093b00e4735b20624410403d4c9b63ddea08329915ccd01d8fddb8eaa8d14f857bf8636c42f8aead65ebd6607035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000224122020302030203020307015471dd1027069078091b3ca48093b00e4735b206244101020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000004410407005471dd1027069078091b3ca48093b00e4735b20624443b9aca0003613e170ccd3d60210a8a915873304f2916101ef41edd8f8c725dd4087aae9180020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000001410107035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000284128020302030325760a1983e67ccdbd15c8a92f71b5fe07c026c66f5bfe95212a6277456fa3700203020202030201020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000032413207035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000000544deadbeef07035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000144114033b02dcf6bc94dbf5fb529ee8341ee6cc93c414b9b5aabbdf9401d3b454c2f51303e746e2599a46e9f11b56ebaf2bcb6dce3881d0f57e4545b32b8fb60246f96666", root_str: "571f39f6e179b508df74cc7871c6602fda7c67e6a9a31ebecd7837c60b4432eb" };
-
-// TestSMTWitnessRetainList, 100 node
-pub(crate) const SMT_TEST_PAYLOAD_6: TestProtocolInputAndRoot = TestProtocolInputAndRoot { byte_str: "010203020302030203020303ea8973d38f29d87599c34d39126e3c00b58ea2c8303350c2143519cb7605f5bf020302030202039c7daba6cb5787b1fa2e1803522ad9c4b2360120563862576df3c7d47857c6a203721e5eaffaf5ca1d7ad83307ee8fe487d3f56e127aad9d0652871cb6897d1ba407035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000050415003117fce402f21f397577181bb77ac70c47b141306d9c8c911a9ec6d94d369a55d0203020302030203020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000095419507035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000091419107035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000013411307035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000934193038f3fa61bcb330bbfc16170074f683d9fa491e10f633a87738930a15a4e986ebe031dbd303cfb87d8fa03bd57c28c3450ebad4b375247316a8b0331c9ee9ce3751e020302030203020302010203020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000027412704440102030407025471dd1027069078091b3ca48093b00e4735b206245820fee777ac8b47f421d9f44ddd09c5efdd8ff00fbc7bbeb8169054d58a1b4b912207035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000058415807035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000064106034def3be799d4e126e6d89fd0dd5d6111ff3a6d5d2aa088dcdb18047b1fd0f849020303bd7f7c3217e47e0cd3dac44c8ba13de803785400a5f284c1231b8344d5a4f1db020307035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000314131036311f89149d48b4e97cda99d7377396c15156dd7ddd3ee0f8eca56902e94ad7002030203020307045471dd1027069078091b3ca48093b00e4735b20624410403d4c9b63ddea08329915ccd01d8fddb8eaa8d14f857bf8636c42f8aead65ebd6607035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000022412203ea19a9b6682b4d98ab419e35b5387d37b1b25731a2b744371077dd6c3a2e93ae02030203020302030202020307015471dd1027069078091b3ca48093b00e4735b20624410107035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000824182020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000004410407005471dd1027069078091b3ca48093b00e4735b20624443b9aca0003fbb670aadaa6f12a2f0849851ee3beead591facafc494c30984f2b910d826e5102030317dd23ee5b9cae7aa5b3f4afb7e62e3e95ed5b8af0d5f8bc7c904ea7b0472065020303bef26264a651a7368eae3074cacd9ad058e22f60d02632603a5d8f26af5c4d6d0202036328ff332cfbf52505cb073de64873642b865bf6bcafceceffaec5a4bad83cec02030203039f5940ca975270b75b29e9dfc4415e150d237dd5d32bd144e4df79518546d1e50203020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000059415902030203020307035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000324132020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000080418007035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000000544deadbeef07035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000094419407035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000144114033b02dcf6bc94dbf5fb529ee8341ee6cc93c414b9b5aabbdf9401d3b454c2f51303842a613fb538b549b8cc7bf73fe0c3647e4323406850667437da7c052ce948c3", root_str: "1d670fb009a8c3c92e3bf1b66cd5cff44aed1967ee1c693faa28fcdfba5c5380" };
-
-// TestSMTWitnessRetainListEmptyVal, 1 node
-pub(crate) const SMT_TEST_PAYLOAD_7: TestProtocolInputAndRoot = TestProtocolInputAndRoot { byte_str: "0102030202020304440102030407025471dd1027069078091b3ca48093b00e4735b206245820fee777ac8b47f421d9f44ddd09c5efdd8ff00fbc7bbeb8169054d58a1b4b912207045471dd1027069078091b3ca48093b00e4735b206244104020307005471dd1027069078091b3ca48093b00e4735b20624443b9aca0007035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000000544deadbeef", root_str: "ed38b6486cbab1828cfe4fa3b2e7907a55fcf67d4a368bfdb65cb7fb7cdce19a" };
-
-// TestSMTWitnessRetainListEmptyVal, 50 node
-pub(crate) const SMT_TEST_PAYLOAD_8: TestProtocolInputAndRoot = TestProtocolInputAndRoot { byte_str: "010203020302030203020303fd6c3576caff9e650a709df487122958032121a9324f245d0588076e7d79ebca0201020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000015411507035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000474147034a0a8ade2d571c3bbafb46ebeaba9af81924e413fe1df89f6ede57203e5516c30203020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000013411307035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000384138035e922f930d52accaba7c46d0a97b2b309290d5385380bf8846acfb38aff55a4d020302030203020302010201020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000027412704440102030407025471dd1027069078091b3ca48093b00e4735b206245820fee777ac8b47f421d9f44ddd09c5efdd8ff00fbc7bbeb8169054d58a1b4b912207035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000000641060314bd622731b8bc10310784233b5d2549b4ba5bd6c70df619ac517314f3526bfb020303ebd30c9db62098e85b738370e9dfd0a7acba5fb656d6b5be4201f8dfbaaebe49020307035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000314131036311f89149d48b4e97cda99d7377396c15156dd7ddd3ee0f8eca56902e94ad7002010203020307045471dd1027069078091b3ca48093b00e4735b20624410403d4c9b63ddea08329915ccd01d8fddb8eaa8d14f857bf8636c42f8aead65ebd6607035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000002241220203020302030202020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000004410407005471dd1027069078091b3ca48093b00e4735b20624443b9aca0003613e170ccd3d60210a8a915873304f2916101ef41edd8f8c725dd4087aae9180020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000001410107035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000284128020302030325760a1983e67ccdbd15c8a92f71b5fe07c026c66f5bfe95212a6277456fa3700203020202030201020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000032413207035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000000544deadbeef07035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000144114033b02dcf6bc94dbf5fb529ee8341ee6cc93c414b9b5aabbdf9401d3b454c2f51303e746e2599a46e9f11b56ebaf2bcb6dce3881d0f57e4545b32b8fb60246f96666", root_str: "73c801adc26f6d9798e7539a3c43abcda94a950171876a85eff1bd306810a9f7" };
-
-// TestSMTWitnessRetainListEmptyVal, 100 node
-pub(crate) const SMT_TEST_PAYLOAD_9: TestProtocolInputAndRoot = TestProtocolInputAndRoot { byte_str: "010203020302030203020303ea8973d38f29d87599c34d39126e3c00b58ea2c8303350c2143519cb7605f5bf020302030202039c7daba6cb5787b1fa2e1803522ad9c4b2360120563862576df3c7d47857c6a203721e5eaffaf5ca1d7ad83307ee8fe487d3f56e127aad9d0652871cb6897d1ba407035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000050415003117fce402f21f397577181bb77ac70c47b141306d9c8c911a9ec6d94d369a55d0203020302030203020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000095419507035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000091419107035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000013411307035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000934193038f3fa61bcb330bbfc16170074f683d9fa491e10f633a87738930a15a4e986ebe031dbd303cfb87d8fa03bd57c28c3450ebad4b375247316a8b0331c9ee9ce3751e020302030203020302010203020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000027412704440102030407025471dd1027069078091b3ca48093b00e4735b206245820fee777ac8b47f421d9f44ddd09c5efdd8ff00fbc7bbeb8169054d58a1b4b912207035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000058415807035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000064106034def3be799d4e126e6d89fd0dd5d6111ff3a6d5d2aa088dcdb18047b1fd0f849020303bd7f7c3217e47e0cd3dac44c8ba13de803785400a5f284c1231b8344d5a4f1db020307035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000314131036311f89149d48b4e97cda99d7377396c15156dd7ddd3ee0f8eca56902e94ad7002030203020307045471dd1027069078091b3ca48093b00e4735b20624410403d4c9b63ddea08329915ccd01d8fddb8eaa8d14f857bf8636c42f8aead65ebd6607035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000022412203ea19a9b6682b4d98ab419e35b5387d37b1b25731a2b744371077dd6c3a2e93ae020302030203020307035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000824182020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000004410407005471dd1027069078091b3ca48093b00e4735b20624443b9aca0003fbb670aadaa6f12a2f0849851ee3beead591facafc494c30984f2b910d826e5102030317dd23ee5b9cae7aa5b3f4afb7e62e3e95ed5b8af0d5f8bc7c904ea7b0472065020303bef26264a651a7368eae3074cacd9ad058e22f60d02632603a5d8f26af5c4d6d0202036328ff332cfbf52505cb073de64873642b865bf6bcafceceffaec5a4bad83cec02030203039f5940ca975270b75b29e9dfc4415e150d237dd5d32bd144e4df79518546d1e50203020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000059415902030203020307035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000324132020307035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000080418007035471dd1027069078091b3ca48093b00e4735b206245820000000000000000000000000000000000000000000000000000000000000000544deadbeef07035471dd1027069078091b3ca48093b00e4735b2062458200000000000000000000000000000000000000000000000000000000000000094419407035471dd1027069078091b3ca48093b00e4735b20624582000000000000000000000000000000000000000000000000000000000000000144114033b02dcf6bc94dbf5fb529ee8341ee6cc93c414b9b5aabbdf9401d3b454c2f51303842a613fb538b549b8cc7bf73fe0c3647e4323406850667437da7c052ce948c3", root_str: "232a950915a7282acf207fb11ad337f4d6af1c3e0ee5452d9c10b8e099a756e4" };
-
-pub(crate) const SMT_TEST_PAYLOAD_10: TestProtocolInputAndRoot = TestProtocolInputAndRoot {
-    byte_str: include_str!("large_test_payloads/smt_test_payload_10.txt"),
-    root_str: "91dfcdeb628dfdc51f3a2ee38cb17c78581e4e7ff91bcc2e327d24a9dfa46982",
-};
-
-pub(crate) const SMT_TEST_PAYLOAD_11: TestProtocolInputAndRoot = TestProtocolInputAndRoot {
-    byte_str: include_str!("large_test_payloads/smt_test_payload_11.txt"),
-    root_str: "e3183936f83466ac3a7c4a0c58bcc62eee4a0a9edb0cbb8bb2ecc3094940f389",
-};
-
-pub(crate) const SMT_TEST_PAYLOAD_12: TestProtocolInputAndRoot = TestProtocolInputAndRoot {
-    byte_str: include_str!("large_test_payloads/smt_test_payload_12.txt"),
-    root_str: "355baf018ce42aae0df4605ff6274da1c2cd650bee756412dd12e60667794049",
-};
-
-pub(crate) const SMT_TEST_PAYLOAD_13: TestProtocolInputAndRoot = TestProtocolInputAndRoot {
-    byte_str: include_str!("large_test_payloads/smt_test_payload_13.txt"),
-    root_str: "91dfcdeb628dfdc51f3a2ee38cb17c78581e4e7ff91bcc2e327d24a9dfa46982",
-};
-
-pub(crate) const SMT_TEST_PAYLOAD_14: TestProtocolInputAndRoot = TestProtocolInputAndRoot {
-    byte_str: include_str!("large_test_payloads/smt_test_payload_14.txt"),
-    root_str: "e3183936f83466ac3a7c4a0c58bcc62eee4a0a9edb0cbb8bb2ecc3094940f389",
-};
-
-pub(crate) const SMT_TEST_PAYLOAD_15: TestProtocolInputAndRoot = TestProtocolInputAndRoot {
-    byte_str: include_str!("large_test_payloads/smt_test_payload_15.txt"),
-    root_str: "355baf018ce42aae0df4605ff6274da1c2cd650bee756412dd12e60667794049",
-};
-
-// 2407440
-pub(crate) const SMT_TEST_PAYLOAD_16: TestProtocolInputAndRoot = TestProtocolInputAndRoot {
-    byte_str: include_str!("large_test_payloads/smt_test_payload_16.txt"),
-    root_str: "76aea30744667aa3cd19c4334dc2712faad7b2f172bac7b7c28dfcbebc478ebd",
-};
-
-// 2407441
-pub(crate) const SMT_TEST_PAYLOAD_17: TestProtocolInputAndRoot = TestProtocolInputAndRoot {
-    byte_str: include_str!("large_test_payloads/smt_test_payload_17.txt"),
-    root_str: "92f67ab1267a150e4cffa902b1c320cf981faeb912f6555238ef9b74eb239740",
-};
-
-// 2407442
-pub(crate) const SMT_TEST_PAYLOAD_18: TestProtocolInputAndRoot = TestProtocolInputAndRoot {
-    byte_str: include_str!("large_test_payloads/smt_test_payload_18.txt"),
-    root_str: "0db217a91709087ad310506b8cc3aa50792b64aac30fd6b25b8c16e6737ae903",
-};
-
-// 2407443
-pub(crate) const SMT_TEST_PAYLOAD_19: TestProtocolInputAndRoot = TestProtocolInputAndRoot {
-    byte_str: include_str!("large_test_payloads/smt_test_payload_19.txt"),
-    root_str: "6bd5c187eedbc3243daa823c5388eb1cf07bd0ed4145a0b093f246f8999b257d",
-};
-
-// 2407444
-pub(crate) const SMT_TEST_PAYLOAD_20: TestProtocolInputAndRoot = TestProtocolInputAndRoot {
-    byte_str: include_str!("large_test_payloads/smt_test_payload_20.txt"),
-    root_str: "c881e24833744cc8c627e59b3541da7d3b6b63c869b50a0978e21d0eea0e436f",
-};
-
-// 2407445
-pub(crate) const SMT_TEST_PAYLOAD_21: TestProtocolInputAndRoot = TestProtocolInputAndRoot {
-    byte_str: include_str!("large_test_payloads/smt_test_payload_21.txt"),
-    root_str: "b39c00ddb244617afcc4b11696ad8dbccad125cceb820278d7e75a5074f4e32f",
-};
-
-// 2407446
-pub(crate) const SMT_TEST_PAYLOAD_22: TestProtocolInputAndRoot = TestProtocolInputAndRoot {
-    byte_str: include_str!("large_test_payloads/smt_test_payload_22.txt"),
-    root_str: "659335a0a85a3337f48591ae6e0494d09352d84a7f76f788ad81f20881bbb598",
-};
-
-// 2407447
-pub(crate) const SMT_TEST_PAYLOAD_23: TestProtocolInputAndRoot = TestProtocolInputAndRoot {
-    byte_str: include_str!("large_test_payloads/smt_test_payload_23.txt"),
-    root_str: "787fb7e084b8d93a77147c837c6ef6e8688a550d7a05db28a3cddeb078049f0a",
-};
-
-// 2407448
-pub(crate) const SMT_TEST_PAYLOAD_24: TestProtocolInputAndRoot = TestProtocolInputAndRoot {
-    byte_str: include_str!("large_test_payloads/smt_test_payload_24.txt"),
-    root_str: "fe8ac2c08a3cec9909fc975534a68479807ce6e617fff2103051ad60b4b52f15",
-};
-
-// 2407449
-pub(crate) const SMT_TEST_PAYLOAD_25: TestProtocolInputAndRoot = TestProtocolInputAndRoot {
-    byte_str: include_str!("large_test_payloads/smt_test_payload_25.txt"),
-    root_str: "24047e78097271a4a9d0e3cf69dd9f5273dc050ae866f2bc5c9450b6f99fe049",
-};
-
-type ProcessMptCompactPrestateFn = ProcessedCompactPrestateFn<Vec<u8>, StateTrieExtractionOutput>;
-
-type ProcessSmtCompactPrestateFn =
-    ProcessedCompactPrestateFn<Vec<u8>, SmtStateTrieExtractionOutput>;
-
-type ProcessedCompactPrestateFn<T, U> = fn(T) -> CompactParsingResult<ProcessedCompactOutput<U>>;
-
-pub(crate) struct TestProtocolInputAndRoot {
-    pub(crate) byte_str: &'static str,
-    pub(crate) root_str: &'static str,
+pub(super) struct TestProtocolInputAndRoot {
+    pub(super) byte_str: &'static str,
+    pub(super) root_str: &'static str,
 }
 
 impl TestProtocolInputAndRoot {
-    #[allow(dead_code)]
-    pub(crate) fn parse_and_check_hash_matches(self) {
-        self.parse_and_check_mpt_trie(MptPreImageProcessing::process_image);
-    }
-
-    pub(crate) fn parse_and_check_hash_matches_with_debug(self) {
-        self.parse_and_check_mpt_trie(MptPreImageProcessing::process_image_debug);
-    }
-
-    pub(crate) fn parse_and_check_hash_matches_with_debug_smt(self) {
-        self.parse_and_check_smt_trie(SmtPreImageProcessing::process_image_debug)
-    }
-
-    fn parse_and_check_mpt_trie(self, process_compact_prestate_f: ProcessMptCompactPrestateFn) {
-        let protocol_bytes = hex::decode(self.byte_str).unwrap();
-
-        let out = match process_compact_prestate_f(protocol_bytes) {
-            Ok(x) => x,
-            Err(err) => panic!("{}", err.to_string()),
-        };
-
-        print_value_and_hash_nodes_of_trie(&out.witness_out.state_trie);
-
-        for (hashed_addr, s_trie) in out.witness_out.storage_tries.iter() {
-            print_value_and_hash_nodes_of_storage_trie(hashed_addr, s_trie);
-        }
-
-        let hash = out.witness_out.state_trie.hash();
-        self.header_and_hash_checks(hash, out.header);
-        Self::assert_non_all_storage_roots_exist_in_storage_trie_map(&out.witness_out);
-    }
-
-    fn parse_and_check_smt_trie(self, process_compact_prestate_f: ProcessSmtCompactPrestateFn) {
-        let protocol_bytes = hex::decode(self.byte_str).unwrap();
-
-        let out: ProcessedCompactOutput<SmtStateTrieExtractionOutput> =
-            process_compact_prestate_f(protocol_bytes).unwrap_or_else(|err| panic!("{}", err));
-        let mut buf: [u8; 32] = [0; 32];
-        hashout2u(out.witness_out.state_trie.root).to_big_endian(&mut buf);
-
-        self.header_and_hash_checks(H256::from_slice(&buf), out.header);
-    }
-
-    fn header_and_hash_checks(self, calculated_hash: TrieRootHash, header: Header) {
+    pub(super) fn header_and_hash_checks(self, calculated_hash: TrieRootHash, header: Header) {
         let expected_hash = TrieRootHash::from_slice(&hex::decode(self.root_str).unwrap());
 
         assert!(header.version_is_compatible(1));
         assert_eq!(calculated_hash, expected_hash);
     }
+}
 
-    fn assert_non_all_storage_roots_exist_in_storage_trie_map(out: &StateTrieExtractionOutput) {
-        let non_empty_account_s_roots = out
-            .state_trie
-            .items()
-            .filter_map(|(addr, data)| {
-                data.as_val().map(|data| {
-                    (
-                        HashedAccountAddr::from_slice(&addr.bytes_be()),
-                        rlp::decode::<AccountRlp>(data).unwrap().storage_root,
-                    )
-                })
-            })
-            .filter(|(_, s_root)| *s_root != EMPTY_TRIE_HASH)
-            .map(|(addr, _)| addr);
+#[cfg(test)]
+pub(super) fn init() {
+    let _ = pretty_env_logger::try_init();
+}
 
-        for account_with_non_empty_root in non_empty_account_s_roots {
-            assert!(out.storage_tries.contains_key(&account_with_non_empty_root));
-        }
-    }
+#[cfg(test)]
+pub(super) fn h_decode_key(h_bytes: &str) -> mpt_trie::nibbles::Nibbles {
+    let bytes = hex::decode(h_bytes).unwrap();
+    key_bytes_to_nibbles(&bytes)
+}
+
+#[cfg(test)]
+pub(super) fn h_decode(b_str: &str) -> Vec<u8> {
+    hex::decode(b_str).unwrap()
 }
