+++ trace_decoder/src/compact/mod.rs
@@ -7,9 +7,15 @@ cfg_if! {
     if #[cfg(feature = "mpt")] {
         pub mod compact_mpt_processing;
         pub mod compact_to_mpt_trie;
+
+        #[cfg(test)]
+        mod complex_test_payloads_mpt;
     } else if #[cfg(feature = "smt")] {
         pub mod compact_smt_processing;
         pub mod compact_to_smt_trie;
+
+        #[cfg(test)]
+        mod complex_test_payloads_smt;
     }
 }
